+++ packages/contracts/contracts/SubtreeUpdateVerifier.sol
@@ -12,26 +12,19 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.5;
-import {ISubtreeUpdateVerifier} from "./interfaces/ISubtreeUpdateVerifier.sol";
 import {Pairing} from "./libs/Pairing.sol";
+import {Groth16} from "./libs/Groth16.sol";
+import {ISubtreeUpdateVerifier} from "./interfaces/ISubtreeUpdateVerifier.sol";
 
 contract SubtreeUpdateVerifier is ISubtreeUpdateVerifier {
     using Pairing for *;
-    struct VerifyingKey {
-        Pairing.G1Point alfa1;
-        Pairing.G2Point beta2;
-        Pairing.G2Point gamma2;
-        Pairing.G2Point delta2;
-        Pairing.G1Point[] IC;
-    }
-    struct Proof {
-        Pairing.G1Point A;
-        Pairing.G2Point B;
-        Pairing.G1Point C;
-    }
 
-    function verifyingKey() internal pure returns (VerifyingKey memory vk) {
-        vk.alfa1 = Pairing.G1Point(
+    function verifyingKey()
+        internal
+        pure
+        returns (Groth16.VerifyingKey memory vk)
+    {
+        vk.alpha1 = Pairing.G1Point(
             20491192805390485299153009773594534940189261866228447918068658471970481763042,
             9383485363053290200918347156157836566562967994039712273449902621266178545958
         );
@@ -58,12 +51,12 @@ contract SubtreeUpdateVerifier is ISubtreeUpdateVerifier {
         );
         vk.delta2 = Pairing.G2Point(
             [
-                11058274375939673914135651655973148518292938769896078273490559643684699521295,
-                20756119478359134777427669301054003966528457459501729206791797430347569979368
+                20982740898175791991495983987857085770426606771988608689973536200514960704278,
+                16230403929903199002213691148035795871020286437662046814473513458528330080854
             ],
             [
-                15359251180066783873261974948214906211946355314268171330918342415682583965671,
-                7857149754711395033794971646937494251878049406928919646595222852330490103495
+                10944084562148035678313884880730141058194572693711617446883569919384462138217,
+                16812161972733117025321508482363121699342797998240361248267230558679611286097
             ]
         );
         vk.IC = new Pairing.G1Point[](5);
@@ -94,60 +87,19 @@ contract SubtreeUpdateVerifier is ISubtreeUpdateVerifier {
         );
     }
 
-    function verify(
-        uint[] memory input,
-        Proof memory proof
-    ) internal view returns (uint) {
-        uint256 snark_scalar_field = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
-        VerifyingKey memory vk = verifyingKey();
-        require(input.length + 1 == vk.IC.length, "verifier-bad-input");
-        // Compute the linear combination vk_x
-        Pairing.G1Point memory vk_x = Pairing.G1Point(0, 0);
-        for (uint i = 0; i < input.length; i++) {
-            require(
-                input[i] < snark_scalar_field,
-                "verifier-gte-snark-scalar-field"
-            );
-            vk_x = Pairing.addition(
-                vk_x,
-                Pairing.scalar_mul(vk.IC[i + 1], input[i])
-            );
-        }
-        vk_x = Pairing.addition(vk_x, vk.IC[0]);
-        if (
-            !Pairing.pairingProd4(
-                Pairing.negate(proof.A),
-                proof.B,
-                vk.alfa1,
-                vk.beta2,
-                vk_x,
-                vk.gamma2,
-                proof.C,
-                vk.delta2
-            )
-        ) return 1;
-        return 0;
-    }
-
     /// @return r  bool true if proof is valid
     function verifyProof(
-        uint[2] memory a,
-        uint[2][2] memory b,
-        uint[2] memory c,
-        uint[4] memory input
+        Groth16.Proof memory proof,
+        uint256[] memory pi
     ) public view override returns (bool r) {
-        Proof memory proof;
-        proof.A = Pairing.G1Point(a[0], a[1]);
-        proof.B = Pairing.G2Point([b[0][0], b[0][1]], [b[1][0], b[1][1]]);
-        proof.C = Pairing.G1Point(c[0], c[1]);
-        uint[] memory inputValues = new uint[](input.length);
-        for (uint i = 0; i < input.length; i++) {
-            inputValues[i] = input[i];
-        }
-        if (verify(inputValues, proof) == 0) {
-            return true;
-        } else {
-            return false;
-        }
+        return Groth16.verifyProof(verifyingKey(), proof, pi);
+    }
+
+    /// @return r bool true if proofs are valid
+    function batchVerifyProofs(
+        Groth16.Proof[] memory proofs,
+        uint256[][] memory allPis
+    ) public view override returns (bool) {
+        return Groth16.batchVerifyProofs(verifyingKey(), proofs, allPis);
     }
 }