+++ packages/contracts/contracts/libs/Utils.sol
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 pragma solidity ^0.8.5;
 import {IWallet} from "../interfaces/IWallet.sol";
+import {Groth16} from "../libs/Groth16.sol";
+import {Pairing} from "../libs/Pairing.sol";
 
 // helpers for converting to/from field elems, uint256s, and/or bytes, and hashing them
 library Utils {
@@ -42,5 +44,16 @@ library Utils {
         elems[4] = note.id;
         elems[5] = note.value;
         return uint256(sha256FieldElems(elems));
+    }
+
+    function proof8ToStruct(
+        uint256[8] memory proof
+    ) internal pure returns (Groth16.Proof memory) {
+        return
+            Groth16.Proof(
+                Pairing.G1Point(proof[0], proof[1]),
+                Pairing.G2Point([proof[2], proof[3]], [proof[4], proof[5]]),
+                Pairing.G1Point(proof[6], proof[7])
+            );
     }
 }