+++ packages/contracts/contracts/test/JoinSplitVerifier.t.sol
@@ -8,55 +8,64 @@ import "@openzeppelin/contracts/utils/Strings.sol";
 
 import {JsonDecodings, JoinSplitProofWithPublicSignals} from "./utils/JsonDecodings.sol";
 import {TestUtils} from "./utils/TestUtils.sol";
-import {IJoinSplitVerifier} from "../interfaces/IJoinSplitVerifier.sol";
 import {JoinSplitVerifier} from "../JoinSplitVerifier.sol";
+import {IJoinSplitVerifier} from "../interfaces/IJoinSplitVerifier.sol";
+import {Groth16} from "../libs/Groth16.sol";
+import {Utils} from "../libs/Utils.sol";
 
 contract TestJoinSplitVerifier is Test, TestUtils, JsonDecodings {
     using stdJson for string;
 
     string constant BASIC_FIXTURE_PATH = "/fixtures/joinsplitProof.json";
+    uint256 constant NUM_PROOFS = 8;
+    uint256 constant NUM_PIS = 9;
 
-    IJoinSplitVerifier verifier;
+    IJoinSplitVerifier joinSplitVerifier;
 
     function setUp() public virtual {
-        verifier = IJoinSplitVerifier(new JoinSplitVerifier());
+        joinSplitVerifier = IJoinSplitVerifier(new JoinSplitVerifier());
+    }
+
+    function loadJoinSplitProof(
+        string memory path
+    ) internal returns (Groth16.Proof memory proof, uint256[] memory pis) {
+        JoinSplitProofWithPublicSignals
+            memory proofWithPIs = loadJoinSplitProofFromFixture(path);
+        proof = Utils.proof8ToStruct(baseProofTo8(proofWithPIs.proof));
+        pis = new uint256[](NUM_PIS);
+        for (uint256 i = 0; i < NUM_PIS; i++) {
+            pis[i] = proofWithPIs.publicSignals[i];
+        }
+
+        return (proof, pis);
     }
 
     function verifyFixture(string memory path) public {
-        JoinSplitProofWithPublicSignals
-            memory proof = loadJoinSplitProofFromFixture(path);
+        (Groth16.Proof memory proof, uint256[] memory pis) = loadJoinSplitProof(
+            path
+        );
+
+        require(joinSplitVerifier.verifyProof(proof, pis), "Invalid proof");
+    }
+
+    function batchVerifyFixture(string memory path) public {
+        Groth16.Proof[] memory proofs = new Groth16.Proof[](NUM_PROOFS);
+        uint256[][] memory pis = new uint256[][](NUM_PROOFS);
+        for (uint256 i = 0; i < NUM_PROOFS; i++) {
+            (proofs[i], pis[i]) = loadJoinSplitProof(path);
+        }
 
         require(
-            verifier.verifyProof(
-                [parseInt(proof.proof.pi_a[0]), parseInt(proof.proof.pi_a[1])],
-                [
-                    [
-                        parseInt(proof.proof.pi_b[0][1]),
-                        parseInt(proof.proof.pi_b[0][0])
-                    ],
-                    [
-                        parseInt(proof.proof.pi_b[1][1]),
-                        parseInt(proof.proof.pi_b[1][0])
-                    ]
-                ],
-                [parseInt(proof.proof.pi_c[0]), parseInt(proof.proof.pi_c[1])],
-                [
-                    proof.publicSignals[0],
-                    proof.publicSignals[1],
-                    proof.publicSignals[2],
-                    proof.publicSignals[3],
-                    proof.publicSignals[4],
-                    proof.publicSignals[5],
-                    proof.publicSignals[6],
-                    proof.publicSignals[7],
-                    proof.publicSignals[8]
-                ]
-            ),
+            joinSplitVerifier.batchVerifyProofs(proofs, pis),
             "Invalid proof"
         );
     }
 
     function testBasicVerify() public {
         verifyFixture(BASIC_FIXTURE_PATH);
+    }
+
+    function testBatchVerify() public {
+        batchVerifyFixture(BASIC_FIXTURE_PATH);
     }
 }