+++ packages/contracts/contracts/test/SubtreeUpdateVerifier.t.sol
@@ -8,50 +8,71 @@ import "@openzeppelin/contracts/utils/Strings.sol";
 
 import {JsonDecodings, SubtreeUpdateProofWithPublicSignals} from "./utils/JsonDecodings.sol";
 import {TestUtils} from "./utils/TestUtils.sol";
+import {Utils} from "../libs/Utils.sol";
 import {ISubtreeUpdateVerifier} from "../interfaces/ISubtreeUpdateVerifier.sol";
+import {Groth16} from "../libs/Groth16.sol";
 import {SubtreeUpdateVerifier} from "../SubtreeUpdateVerifier.sol";
 
 contract TestSubtreeUpdateVerifier is Test, TestUtils, JsonDecodings {
     using stdJson for string;
 
     string constant BASIC_FIXTURE_PATH = "/fixtures/subtreeupdateProof.json";
+    uint256 constant NUM_PROOFS = 8;
+    uint256 constant NUM_PIS = 4;
 
-    ISubtreeUpdateVerifier verifier;
+    ISubtreeUpdateVerifier subtreeUpdateVerifier;
 
     function setUp() public virtual {
-        verifier = ISubtreeUpdateVerifier(new SubtreeUpdateVerifier());
+        subtreeUpdateVerifier = ISubtreeUpdateVerifier(
+            new SubtreeUpdateVerifier()
+        );
+    }
+
+    function loadSubtreeUpdateProof(
+        string memory path
+    ) internal returns (Groth16.Proof memory proof, uint256[] memory pis) {
+        SubtreeUpdateProofWithPublicSignals
+            memory proofWithPIs = loadSubtreeUpdateProofFromFixture(path);
+        proof = Utils.proof8ToStruct(baseProofTo8(proofWithPIs.proof));
+        pis = new uint256[](NUM_PIS);
+        for (uint256 i = 0; i < NUM_PIS; i++) {
+            pis[i] = proofWithPIs.publicSignals[i];
+        }
+
+        return (proof, pis);
     }
 
     function verifyFixture(string memory path) public {
-        SubtreeUpdateProofWithPublicSignals
-            memory proof = loadSubtreeUpdateProofFromFixture(path);
+        (
+            Groth16.Proof memory proof,
+            uint[] memory pis
+        ) = loadSubtreeUpdateProof(path);
+        require(subtreeUpdateVerifier.verifyProof(proof, pis), "Invalid proof");
+    }
+
+    function batchVerifyFixture(string memory path) public {
+        Groth16.Proof[] memory proofs = new Groth16.Proof[](NUM_PROOFS);
+        uint[][] memory pis = new uint256[][](NUM_PROOFS);
+        for (uint256 i = 0; i < NUM_PROOFS; i++) {
+            (
+                Groth16.Proof memory proof,
+                uint[] memory _pis
+            ) = loadSubtreeUpdateProof(path);
+            proofs[i] = proof;
+            pis[i] = _pis;
+        }
 
         require(
-            verifier.verifyProof(
-                [parseInt(proof.proof.pi_a[0]), parseInt(proof.proof.pi_a[1])],
-                [
-                    [
-                        parseInt(proof.proof.pi_b[0][1]),
-                        parseInt(proof.proof.pi_b[0][0])
-                    ],
-                    [
-                        parseInt(proof.proof.pi_b[1][1]),
-                        parseInt(proof.proof.pi_b[1][0])
-                    ]
-                ],
-                [parseInt(proof.proof.pi_c[0]), parseInt(proof.proof.pi_c[1])],
-                [
-                    proof.publicSignals[0],
-                    proof.publicSignals[1],
-                    proof.publicSignals[2],
-                    proof.publicSignals[3]
-                ]
-            ),
+            subtreeUpdateVerifier.batchVerifyProofs(proofs, pis),
             "Invalid proof"
         );
     }
 
     function testBasicVerify() public {
         verifyFixture(BASIC_FIXTURE_PATH);
+    }
+
+    function testBatchVerify() public {
+        batchVerifyFixture(BASIC_FIXTURE_PATH);
     }
 }