+++ packages/contracts/contracts/test/utils/TestJoinSplitVerifier.sol
@@ -1,15 +1,21 @@
 //SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.5;
 
-import "../../interfaces/IJoinSplitVerifier.sol";
 import "./Pairing.sol";
+import {IJoinSplitVerifier} from "../../interfaces/IJoinSplitVerifier.sol";
+import {Groth16} from "../../libs/Groth16.sol";
 
 contract TestJoinSplitVerifier is IJoinSplitVerifier {
     function verifyProof(
-        uint256[2] memory a,
-        uint256[2][2] memory b,
-        uint256[2] memory c,
-        uint256[9] memory input
+        Groth16.Proof memory proof,
+        uint256[] memory pis
+    ) external view override returns (bool) {
+        return true;
+    }
+
+    function batchVerifyProofs(
+        Groth16.Proof[] memory proofs,
+        uint256[][] memory pis
     ) external view override returns (bool) {
         return true;
     }