+++ packages/contracts/contracts/test/utils/TestSubtreeUpdateVerifier.sol
@@ -1,15 +1,21 @@
 //SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.5;
 
-import "../../interfaces/ISubtreeUpdateVerifier.sol";
 import "./Pairing.sol";
+import {ISubtreeUpdateVerifier} from "../../interfaces/ISubtreeUpdateVerifier.sol";
+import {Groth16} from "../../libs/Groth16.sol";
 
 contract TestSubtreeUpdateVerifier is ISubtreeUpdateVerifier {
     function verifyProof(
-        uint256[2] memory a,
-        uint256[2][2] memory b,
-        uint256[2] memory c,
-        uint256[4] memory input
+        Groth16.Proof memory proof,
+        uint256[] memory pis
+    ) external view override returns (bool) {
+        return true;
+    }
+
+    function batchVerifyProofs(
+        Groth16.Proof[] memory proofs,
+        uint256[][] memory pis
     ) external view override returns (bool) {
         return true;
     }