+++ packages/contracts/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ### Unreleased
 
-- Add a `SubtreeUpdate` event to `CommitmentTreeManager` for when a subtree update is committed 
+- Use the batch verifier to batch verify all joinsplit proofs in a bundle
+- Change verifier interfaces to take structs instead of `uint256` arrays
+- Add a `SubtreeUpdate` event to `CommitmentTreeManager` for when a subtree update is committed
+- Add `test:gas-report` command
+- Add batch proof verification
 - Add `test:gas-report` command
 - Migrate for joinsplit
   - Compute operationDigest once for the entire operation